+++ docs/components/LightBox/Web.stories.tsx
@@ -18,11 +18,13 @@ const images = [
     url: "https://images.unsplash.com/photo-1597201278257-3687be27d954?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80",
     caption:
       "This was the view of Bushart Gardens in Victoria, BC, Canada in July from a hill.",
+    alt: "Vibrant garden with colorful flowers, green bushes, and trees, arranged in neat patterns with black garden lights on lush green grass.",
   },
   {
     title: "A house",
     url: "https://images.unsplash.com/photo-1592595896616-c37162298647?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80",
     caption: "House with a garden.",
+    alt: "White brick house with dark gray roof, brown shutters, a small porch with a lantern, and a green lawn with flower beds and young trees.",
   },
   {
     url: "https://images.unsplash.com/photo-1532302780319-95689ab9d79a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=687&q=80",